fixes#20 fixes#91 fixes#11 fixes#21
Removed the "empty.png" beacuse iPhone is slowing when it has to stretch it on every cell.  Instead, cells
have a random choice of grey/blue background-color.  Each role has a corresponding
color and it is used in the guesser and clues page.  When you learn a player's role, their avatar takes on that new BG color.  TODO: makes cubes show bg
color (we need a transparent cubicle to do this).